requirements.txt
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
pytest-cov==4.1.0
beautifulsoup4==4.12.2
tweepy==4.14.0
praw==7.7.1
anthropic==0.7.4
smolagents>=0.1.0
backoff>=2.2.1
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.1